Add a partial template to add a 'Created' and eventually 'Updated' dates at the beginning of a page.
For the moment, this is not used. But to integrate this partial in a template
file, nothing less simple than inserting the following snippet wherever one
wants to add these dates in the final HTML page:

```
{{- partial "page-dates.html" . }}
```

I have also added some simple CSS, but it can obviously be adapted to the IVOA
website.